Move replication-related functions into separate module
