+++ lz/iterating.py
@@ -33,6 +33,7 @@ from .hints import (Domain,
                     Predicate,
                     Range,
                     Sortable)
+from .replication import duplicate
 from .textual import (decoder,
                       read_batch_from_end,
                       split)
@@ -350,71 +351,6 @@ def flatmapper(map_: Map[Domain, Iterable[Range]]
     and flattens results.
     """
     return compose(flatten, mapper(map_))
-
-
-def replicator(count: int) -> Map[Domain, Iterable[Domain]]:
-    """
-    Returns function that replicates passed object.
-    """
-    return functools.partial(replicate,
-                             count=count)
-
-
-@functools.singledispatch
-def replicate(object_: Domain,
-              *,
-              count: int) -> Iterable[Domain]:
-    """
-    Returns given number of object replicas.
-    """
-    raise TypeError('Unsupported object type: {type}.'
-                    .format(type=type(object_)))
-
-
-@replicate.register(object)
-# mappings cannot be replicated as other iterables
-# since they are iterable only by key
-@replicate.register(abc.Mapping)
-# immutable strings represent a special kind of iterables
-# that can be replicated by simply repeating
-@replicate.register(bytes)
-@replicate.register(str)
-def replicate_object(object_: Domain,
-                     *,
-                     count: int) -> Iterable[Domain]:
-    """
-    Returns object repeated given number of times.
-    """
-    return itertools.repeat(object_, count)
-
-
-@replicate.register(abc.Iterable)
-def replicate_iterable(object_: Iterable[Domain],
-                       *,
-                       count: int) -> Iterable[Iterable[Domain]]:
-    """
-    Returns given number of iterable replicas.
-    """
-    iterator = iter(object_)
-    queues = [deque() for _ in itertools.repeat(None, count)]
-
-    def replica(queue: deque) -> Iterable[Domain]:
-        while True:
-            if not queue:
-                try:
-                    element = next(iterator)
-                except StopIteration:
-                    return
-                element_copies = replicate(element,
-                                           count=count)
-                for sub_queue, element_copy in zip(queues, element_copies):
-                    sub_queue.append(element_copy)
-            yield queue.popleft()
-
-    yield from map(replica, queues)
-
-
-duplicate = replicator(2)
 
 
 def header(size: int) -> Operator[Iterable[Domain]]:
