+++ tests/iterating_tests/test_first.py
@@ -1,8 +1,8 @@
 from typing import (Any,
                     Iterable)
 
-from lz.iterating import (duplicate,
-                          first)
+from lz.iterating import first
+from lz.replication import duplicate
 
 
 def test_basic(non_empty_iterable: Iterable[Any]) -> None:
