+++ tests/iterating_tests/test_last.py
@@ -1,10 +1,10 @@
 from typing import (Any,
                     Iterable)
 
-from lz.iterating import (duplicate,
-                          first,
+from lz.iterating import (first,
                           last,
                           reverse)
+from lz.replication import duplicate
 
 
 def test_basic(non_empty_iterable: Iterable[Any]) -> None:
